refactor: Modularize, Types, code, directory structure
