+++ src/mpr/utility/addData.ts
@@ -1,0 +1,16 @@
+import { VarsChild } from '../modules/parseData';
+
+export function addData(
+  varsChild: Partial<VarsChild>,
+  value: string | number | undefined,
+): Partial<VarsChild> {
+  //just to emulate prev code behaviour
+  if (value === undefined) return varsChild;
+
+  if (Array.isArray(varsChild.data)) {
+    varsChild.data.push(value);
+  } else {
+    varsChild.data = [value];
+  }
+  return varsChild;
+}
