+++ src/mpr/utility/pascalString.ts
@@ -1,0 +1,6 @@
+import {IOBuffer} from "iobuffer";
+
+export function pascalString(buffer: IOBuffer): string {
+  const length = buffer.readUint8();
+  return buffer.readChars(length);
+}
